requirements.txt
numpy>=1.19.2
scipy>=1.5.3
torch==1.7.1
torchvision~=0.8.2
scikit-learn~=0.24.0
pillow~=8.0.1
wandb~=0.10.11
sklearn~=0.0
pytest
pytorch_lightning
pytorch-lightning-bolts
gym
pdoc3
office31
plotly
wget
matplotlib